games should not be sorted on alphabetical order of the nameCode (that is meaningless to the user) but on the alphabetical order to the translation of the nameCode in the user language
