+++ gazeplay-games-commons/src/main/java/net/gazeplay/GameSummary.java
@@ -6,17 +6,12 @@ import lombok.Getter;
 import lombok.Singular;
 import lombok.extern.slf4j.Slf4j;
 
-import java.util.Comparator;
-import java.util.HashMap;
-import java.util.Map;
 import java.util.SortedSet;
 
 @Builder
 @Slf4j
 @EqualsAndHashCode
-public class GameSummary implements Comparable<GameSummary> {
-
-    private static final CategoriesCollectionComparator categoriesCollectionComparator = new CategoriesCollectionComparator();
+public class GameSummary {
 
     @Getter
     private final String nameCode;
@@ -50,45 +45,5 @@ public class GameSummary implements Comparable<GameSummary> {
     @Getter
     private final int priority;
 
-    @Override
-    public int compareTo(GameSummary o) {
-        return Comparator
-            .comparing(GameSummary::getCategories, categoriesCollectionComparator)
-            .thenComparing(Comparator.comparing(GameSummary::getPriority).reversed())
-            .thenComparing(GameSummary::getNameCode)
-            .compare(this, o);
-    }
 
-    private static class CategoriesCollectionComparator implements Comparator<SortedSet<GameCategories.Category>> {
-
-        private static final Map<GameCategories.Category, Integer> weightMap = createWeightMap();
-
-        private static Map<GameCategories.Category, Integer> createWeightMap() {
-            GameCategories.Category[] values = GameCategories.Category.values();
-            Map<GameCategories.Category, Integer> result = new HashMap<>();
-            int weight = values.length;
-            for (GameCategories.Category category : values) {
-                result.put(category, (int) Math.pow(weight, 2));
-                weight--;
-            }
-            return result;
-        }
-
-        private static int computeTotalWeight(SortedSet<GameCategories.Category> categories) {
-            int result = 0;
-            for (GameCategories.Category category : categories) {
-                result += weightMap.get(category);
-            }
-            return result;
-        }
-
-        @Override
-        public int compare(SortedSet<GameCategories.Category> o1, SortedSet<GameCategories.Category> o2) {
-            int o1Weight = computeTotalWeight(o1);
-            int o2Weight = computeTotalWeight(o2);
-            int diff = o2Weight - o1Weight;
-            return Integer.compare(diff, 0);
-        }
-
-    }
 }
