+++ gazeplay/src/main/java/net/gazeplay/GamesLocator.java
@@ -1,9 +1,11 @@
 package net.gazeplay;
 
+import net.gazeplay.commons.ui.Translator;
+
 import java.util.List;
 
 public interface GamesLocator {
 
-    List<GameSpec> listGames();
+    List<GameSpec> listGames(Translator translator);
 
 }
