+++ gazeplay/src/main/java/net/gazeplay/gameslocator/CachingGamesLocator.java
@@ -3,6 +3,7 @@ package net.gazeplay.gameslocator;
 import lombok.AllArgsConstructor;
 import net.gazeplay.GameSpec;
 import net.gazeplay.GamesLocator;
+import net.gazeplay.commons.ui.Translator;
 
 import java.util.Collections;
 import java.util.List;
@@ -16,10 +17,10 @@ public class CachingGamesLocator implements GamesLocator {
     private final AtomicReference<List<GameSpec>> cachedResult = new AtomicReference<>();
 
     @Override
-    public List<GameSpec> listGames() {
+    public List<GameSpec> listGames(Translator translator) {
         List<GameSpec> result = cachedResult.get();
         if (result == null) {
-            result = Collections.unmodifiableList(delegate.listGames());
+            result = Collections.unmodifiableList(delegate.listGames(translator));
             cachedResult.set(result);
         }
         return result;
